% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_caged.R
\name{baixa_caged}
\alias{baixa_caged}
\title{Funcao para baixar microdados do Novo Caged}
\usage{
baixa_caged(ano, mes_inicial, mes_final = mes_inicial)
}
\arguments{
\item{ano}{Ano dos arquivos que devem ser baixados. Nao e possivel baixar microdados de meses de anos distintos.}

\item{mes_inicial}{Mes cujos microdados vao ser baixados. No caso de baixar mais de um mes, esse argumento deve ser o mes mais distante.}

\item{mes_final}{Argumento opcional. No caso de baixar varios meses, se refere ao mes mais recente. Se for baixar apenas um mes o preenchimento e opcional.}
}
\description{
Funcao para baixar microdados do Novo Caged
}
