% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_novo_caged.R
\name{baixa_novo_caged}
\alias{baixa_novo_caged}
\title{Funcao para baixar, descomprimir e ler microdados do Novo Caged}
\usage{
baixa_novo_caged(
  ano,
  mes_inicial,
  mes_final = mes_inicial,
  caminho = "atualizacao/caged_atualizacao.parquet",
  apaga7z = T
)
}
\arguments{
\item{ano}{Ano dos arquivos que devem ser baixados. Nao e possivel baixar microdados de meses de anos distintos.}

\item{mes_inicial}{Mes cujos microdados vao ser baixados. No caso de baixar mais de um mes, esse argumento deve ser o mes mais distante.}

\item{mes_final}{Argumento opcional. No caso de baixar varios meses, se refere ao mes mais recente. Se for baixar apenas um mes o preenchimento e opcional.}

\item{caminho}{Caminho contendo nome da pasta e do arquivo em que os dados serao salvos. Deve terminar em ".parquet".}

\item{apaga7z}{Argumento logico. Se TRUE, apaga os arquivos intermeiarios - dos microdados baixados e descompactados em formato .txt. CUIDADO: na pratica, apaga TODOS os arquivos .7z que estiverem salvas na pasta de trabalho!}
}
\description{
Funcao para baixar, descomprimir e ler microdados do Novo Caged
}
\examples{
# library(DieeseCAGED)
# EXEMPLO 1: julho/22
# baixa_novo_caged(7, ano=2022, caminho="atualizacao/caged_202207.parquet")

# # para ler:
# caged_2021m7 <- arrow::read_parquet("atualizacao/caged_202207.parquet")
# # para salvar em csv
# salva_parquet_em_csv(caminho_parquet="atualizacao/caged_202207.parquet",
#                      caminho_csv="atualizacao_csv/caged_202207.csv")
# # para zipar a pasta com csv
# Zipar_em7zip("atualizacao_csv")
}
