% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflator_inpc.R
\name{deflator_inpc}
\alias{deflator_inpc}
\title{Funcao para baixar microdados do Novo Caged}
\usage{
deflator_inpc(base_caged, mes_base = NULL)
}
\arguments{
\item{base_caged}{Base de dados do caged a ser.}

\item{mes_base}{Mes base para o calculo do deflator. Se nulo, calcula a partir do mes mais recete do caged.}
}
\description{
Funcao para baixar microdados do Novo Caged
}
