% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descomprime_bases.R
\name{descomprime_bases}
\alias{descomprime_bases}
\title{Funcao para descompactar os arquivos baixados (que vao para uma pasta chamada PastaMes)}
\usage{
descomprime_bases()
}
\description{
Funcao para descompactar os arquivos baixados (que vao para uma pasta chamada PastaMes)
}
\examples{
novocaged::descomprime_bases()
}
