% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faz_media.R
\name{faz_media}
\alias{faz_media}
\title{Funcao para fazer tabelas de médias no Novo Caged.}
\usage{
faz_media(dados, ..., var_media = salario_real, tira_outliers = T)
}
\arguments{
\item{dados}{nome da base de dados do novo caged. Necessario que contenha as variaveis: saldomovimentacao e indicadordeexclusao}

\item{...}{uma ou mais variaveis que se deseja cruzar}

\item{var_media}{variavel que se deseja tirar media. P. ex.: salario. No default, calcula o salario real com base no INPC do ultimo mes que aparece na base do Caged}

\item{tira_outliers}{tira salarios minimos abaixo de 0,3SM e acima de 150SM?}
}
\description{
Funcao para fazer tabelas de médias no Novo Caged.
}
