% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salva_bases.R
\name{salva_bases}
\alias{salva_bases}
\title{Funcao para salvar os microdados baixados e descompacados (em formato .txt) do Novo Caged}
\usage{
salva_bases(
  ano,
  mes_inicial,
  mes_final = mes_inicial,
  caminho = "atualizacao/caged_atualizacao.parquet"
)
}
\arguments{
\item{ano}{Ano dos arquivos que devem ser baixados. Nao e possivel baixar microdados de meses de anos distintos.}

\item{mes_inicial}{Mes cujos microdados vao ser baixados. No caso de baixar mais de um mes, esse argumento deve ser o mes mais distante.}

\item{mes_final}{Argumento opcional. No caso de baixar varios meses, se refere ao mes mais recente. Se for baixar apenas um mes o preenchimento e opcional.}

\item{caminho}{Caminho contendo nome da pasta e do arquivo em que os dados serao salvos. Deve terminar em ".parquet".}
}
\description{
Funcao para salvar os microdados baixados e descompacados (em formato .txt) do Novo Caged
}
