% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salva_parquet_em_csv.R
\name{salva_parquet_em_csv}
\alias{salva_parquet_em_csv}
\title{Funcao para transformar arquivos .parquet em arquivos .csv}
\usage{
salva_parquet_em_csv(
  caminho_parquet = "atualizacao/caged_atualizacao.parquet",
  caminho_csv = "atualizacao_csv/caged_atualizacao.csv"
)
}
\arguments{
\item{caminho_parquet}{caminho do arquivo em formato .parquet.}

\item{caminho_csv}{caminho em que o arquivo no formato .csv deve ser salvo.}
}
\description{
Funcao para transformar arquivos .parquet em arquivos .csv
}
