% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SM_DIEESE.R
\name{SM_DIEESE}
\alias{SM_DIEESE}
\title{Funcao para baixar e arrumar em formato de tabela a serie historica de salarios minimos e salario minimos necessarios calculados pelo DIEESE}
\usage{
SM_DIEESE()
}
\description{
Funcao para baixar e arrumar em formato de tabela a serie historica de salarios minimos e salario minimos necessarios calculados pelo DIEESE
}
