% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zipar_em7zip.R
\name{Zipar_em7zip}
\alias{Zipar_em7zip}
\title{Funcao para compactar arquivos .csv}
\usage{
Zipar_em7zip(caminho_pasta = "atualizacao_csv")
}
\arguments{
\item{caminho_pasta}{caminho da pasta contendo os arquivos em formato .csv, que deverao ser compactados. O arquivo fica salvo na pasta de trabalho.}
}
\description{
Funcao para compactar arquivos .csv
}
